# Node.js/Tasks_Intermediate/week7/task/JWT_project/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: JWT API
  description: A sample API for managing users and products.
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - status
            - message 
          properties:
            status:
              type: integer
              description: The HTTP status code of the error
              enum: [401, 403,409, 500]
            message:
              type: string
              description: A human-readable error message.
...
    LoginResponse:
      description: Response body containing authentication tokens.
      type: object
      required:
        - accessToken
        - refreshToken
      properties:
        accessToken:
          type: string
          description: JWT Access Token (short-lived)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIxMjMiLCJpYXQiOjE2NzgyOTk3ODQsImV4cCI6MTY3ODI5OTg0NH0.xyzabc
        refreshToken:
          type: string
          description: JWT Refresh Token (long-lived)
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiIxMjMiLCJpYXQiOjE2NzgyOTk3ODQsImV4cCI6MTY3ODkwNDU4NH0.defghi
...
    RefreshTokenRequest:
      description: Request Body for Refresh the accessToken and get new RefreshToken too
      type: object
      required:
      - refreshToken
      properties:
        refreshToken:
          type: string
          description: The refresh token obtained during login or previous refresh
...
  securitySchemes:
    bearerAuth: 
      type: http
      scheme: bearer
      bearerFormat: JWT
  examples:
    InternalServerError:
      summary: Internal Server Error
      value:
        error:
          status: 505
          message: "Internal server error"
...
    UnauthorizedGeneric:
      summary: Generic Unauthorized Response
      value:
        error:
          status: 401
          message: "Unauthorized"
    TokensResponseExample:
      summary: Response of the registration process
      value:
        accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2ODYxNDgyMThmZTY1MjAxYjlmZWIwOGUiLCJyb2xlIjoiYWRtaW4iLCJpYXQiOjE3NTEyMDU5MjIsImV4cCI6MTc1MTIwNjA0MiwiYXVkIjoiNjg2MTQ4MjE4ZmU2NTIwMWI5ZmViMDhlIiwiaXNzIjoiZ29vZ2xlLmNvbSJ9.X5536Mt5khuDCSW3zK9BJxFXs-KAmSL0Bb3uMegD7pw"
        refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE3NTEyMDU5MjIsImV4cCI6MTc4Mjc2MzUyMiwiYXVkIjoiNjg2MTQ4MjE4ZmU2NTIwMWI5ZmViMDhlIiwiaXNzIjoiZ29vZ2xlLmNvbSJ9.4pNTOcOiO-QO8Iv2P3_8X6FzVbBH6vplhLyokQCuzCg"
...
    RefreshTokenBodyExample: 
      summary: Example of the body of the refresh access Token process
      value:
        refreshToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpYXQiOjE3NTEyMjYwMDEsImV4cCI6MTc4Mjc4MzYwMSwiYXVkIjoiNjgxMjljZWE0NGZiMGFjZDMxZTQ1OTZmIiwiaXNzIjoiZ29vZ2xlLmNvbSJ9.zizv5HJGbsCwiKwWKhL_THVnX6uaFgl6dGUL4NC7aGU"
    BadRequestMissingRefreshTokenExample:
      summary: Bad Request - Missing Refresh Token
      value:
        error:
          status: 400
          message: "Bad Request: Refresh token is required."
paths:
...
  /auth/refresh-token:
    post:
      summary: Refresh Access Token
      operationId: refreshToken
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'  
            examples:
              RefreshTokenBodyExample:
                $ref: '#/components/examples/RefreshTokenBodyExample'
      responses:
        200:
          description: successfully Refresh Access Token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                TokensResponseExample:
                  $ref: '#/components/examples/TokensResponseExample'
        401:
          description: "Unauthorized. This typically indicates that authentication credentials (e.g., JWT) are missing, invalid, or expired or the accessToken is overwritten in Redis database.
            The specific reason is not exposed for security purposes. Please ensure a valid and unexpired token is provided or login again."
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                UnauthorizedGeneric:
                  $ref: '#/components/examples/UnauthorizedGeneric'
        400:
          description: Bad Request - Missing refresh token in request body. 
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                BadRequestMissingRefreshTokenExample:
                  $ref: '#/components/examples/BadRequestMissingRefreshTokenExample'
        500:
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                InternalServerError:
                  $ref: '#/components/examples/InternalServerError'
...